Correcting compilation and code generation problems, arranging a segfault that was occurring during compilation, and minor improvements in the module system and the code in general

+++ kernel/core/vfs/interfaces.zig
@@ -57,13 +57,13 @@ pub const InodeOp_T: type = struct {
 };
 
 pub const Inode_T: type = struct {
-    ino: u32, // Numero do inode
+    ino: usize, // Numero do inode
     type: FileType_T, // Tipo do arquivo
     uid: u16, // ID do usuario
     gid: u32, // ID do grupo
     mode: u16, // Permissoes do arquivo
     nlinks: u16, // Quantidade de links que apontam para esse inode
-    data_block: u32, // Aponta para qual bloco inicial estao os dados desse arquivo 
+    data_block: usize, // Aponta para qual bloco inicial estao os dados desse arquivo 
     private: ?*anyopaque, // Dados internos do FS
     ops: ?*InodeOp_T, // Operaçoes para esse inode
 };
@@ -75,11 +75,11 @@ pub const Dentry_T: type = struct {
 
 pub const Superblock_T: type = struct {
     magic: u32, // Flag magica
-    block_size: u32, // Tamanho em bytes de cada bloco do disco
-    total_blocks: u32, // Quantidade total de blocos disponiveis no dispositivo
-    total_inodes: u32, // Numero total de inodes disponiveis
-    inode_table_start: u32, // Offset(em blocos) de onde começa a tabela de inodes
-    data_block_start: u32, // Offset no disco onde começa a area de dados dos arquivos
+    block_size: usize, // Tamanho em bytes de cada bloco do disco
+    total_blocks: usize, // Quantidade total de blocos disponiveis no dispositivo
+    total_inodes: usize, // Numero total de inodes disponiveis
+    inode_table_start: usize, // Offset(em blocos) de onde começa a tabela de inodes
+    data_block_start: usize, // Offset no disco onde começa a area de dados dos arquivos
     root_inode: ?*Inode_T, // Ponteiro para o inode raiz do sistema de arquivos
     private_data: ?*anyopaque, // Dados internos do FS (cast dinamico)
 };
